--- src/main/java/ru/clevertec/core/parser/JsonParser2.java
@@ -1,17 +1,0 @@
-package ru.clevertec.core.parser;
-
-import ru.clevertec.core.node.Node;
-
-public interface JsonParser2 {
-    char ARRAY_START = '[';
-    char ARRAY_END = ']';
-    char OBJECT_START = '{';
-    char OBJECT_END = '}';
-    char QUOTES = '"';
-    char COMMA = ',';
-    char BACKSLASH = '\\';
-
-    int parse(Node node, char[] json, int start);
-    int parse(Node node, char[] json);
-
-}
