+++ src/main/java/ru/clevertec/core/parser/ParserArray.java
@@ -1,10 +1,9 @@
-package ru.clevertec.core.parser.temp;
+package ru.clevertec.core.parser;
 
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import ru.clevertec.core.node.ArrayNode;
 import ru.clevertec.core.node.Node;
+import ru.clevertec.util.ReplacerComma;
 
 @RequiredArgsConstructor
 public class ParserArray implements ParserJson{
