+++ src/main/java/ru/clevertec/core/parser/ParserFactory.java
@@ -1,4 +1,4 @@
-package ru.clevertec.core.parser.temp;
+package ru.clevertec.core.parser;
 
 public interface ParserFactory {
