+++ src/main/java/ru/clevertec/core/parser/ParserValue.java
@@ -1,4 +1,4 @@
-package ru.clevertec.core.parser.temp;
+package ru.clevertec.core.parser;
 
 import ru.clevertec.core.node.Node;
 import ru.clevertec.core.node.ValueNode;
