+++ src/test/java/ru/clevertec/ConverterTest.java
@@ -11,9 +11,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import ru.clevertec.core.AbstractContainer;
-import ru.clevertec.core.parser.Parser2;
-import ru.clevertec.core.node.NodeFactory;
-import ru.clevertec.core.parser.temp.ParserFactoryImpl;
+import ru.clevertec.core.parser.ParserFactoryImpl;
 import ru.clevertec.core.service.deserialization.NodeConverterFactory;
 import ru.clevertec.core.service.serialization.ConverterFactoryImpl;
 import ru.clevertec.domain.Customer;
@@ -22,7 +20,6 @@ import ru.clevertec.domain.InnerObjectsFlower;
 import ru.clevertec.domain.InnerObjectsWithMapAndList;
 import ru.clevertec.domain.TimeFlower;
 import ru.clevertec.domain.UuidFlower;
-import ru.clevertec.util.StringCleanerImpl;
 import ru.clevertec.util.TestHelper;
 
 import java.io.IOException;
