+++ src/test/java/ru/clevertec/core/ParserFactoryImplTest.java
@@ -1,12 +1,12 @@
 package ru.clevertec.core;
 
 import org.junit.jupiter.api.Test;
-import ru.clevertec.core.parser.temp.ParserArray;
-import ru.clevertec.core.parser.temp.ParserFactory;
-import ru.clevertec.core.parser.temp.ParserFactoryImpl;
-import ru.clevertec.core.parser.temp.ParserJson;
-import ru.clevertec.core.parser.temp.ParserOneLineObject;
-import ru.clevertec.core.parser.temp.ParserValue;
+import ru.clevertec.core.parser.ParserArray;
+import ru.clevertec.core.parser.ParserFactory;
+import ru.clevertec.core.parser.ParserFactoryImpl;
+import ru.clevertec.core.parser.ParserJson;
+import ru.clevertec.core.parser.ParserOneLineObject;
+import ru.clevertec.core.parser.ParserValue;
 
 import static org.assertj.core.api.Assertions.assertThat;
