+++ src/test/java/ru/clevertec/core/ReplacerCommaTest.java
@@ -1,7 +1,7 @@
 package ru.clevertec.core;
 
 import org.junit.jupiter.api.Test;
-import ru.clevertec.core.parser.temp.ReplacerComma;
+import ru.clevertec.util.ReplacerComma;
 import ru.clevertec.util.TestHelper;
 
 import java.io.IOException;
